Rename header files to better reflect their content
