+++ protocol/typeSerialization.hpp
